backend_mc/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-editorjs-fields==0.2.7
django-jazzmin==3.0.3
djangorestframework==3.16.1
pillow==12.1.1
pillow_heif==1.3.0
psycopg2==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3

# Dependencias para Producción (Railway / Neon / Cloudflare R2)
django-storages[s3]==1.14.4
boto3==1.36.0
dj-database-url==2.3.0
gunicorn==23.0.0
whitenoise==6.9.0
